update project structure -created new class Validator -deleted validator's entity -deleted class MergeSortStrategy -renamed pachage and added all interfaces in one package
