+++ src/main/java/algorithm/MergeSort.java
@@ -3,7 +3,7 @@
  */
 package algorithm;
 
-import strategy.SortStrategy;
+import interfaces.SortStrategy;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
