--- src/main/java/strategy/MergeSortStrategy.java
@@ -1,4 +1,0 @@
-package strategy;
-
-public class MergeSortStrategy {
-}
