--- src/main/java/validator/BookValidator.java
@@ -1,4 +1,0 @@
-package validator;
-
-public class BookValidator {
-}
