--- src/main/java/validator/CarValidator.java
@@ -1,4 +1,0 @@
-package validator;
-
-public class CarValidator {
-}
