--- src/main/java/validator/RootVegetableValidator.java
@@ -1,4 +1,0 @@
-package validator;
-
-public class RootVegetableValidator {
-}
