+++ src/main/java/validator/Validator.java
@@ -1,0 +1,4 @@
+package validator;
+
+public class Validator {
+}
